Pulling sizecost and completed reset and login

+++ website/public/webfuncs.js
@@ -1,6 +1,5 @@
 const url = 'https://main-zo6hfspdfa-uc.a.run.app/';
 const bearer = 'bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJPbmxpbmUgSldUIEJ1aWxkZXIiLCJpYXQiOjE2ODA2NTg3MDIsImV4cCI6MTcxMjE5NDcyMiwiYXVkIjoid3d3LmV4YW1wbGUuY29tIiwic3ViIjoianJvY2tldEBleGFtcGxlLmNvbSIsIlVzZXJuYW1lIjoicHJpeWFua2EifQ.ln-9bSIm9Br2u2OJBb5Cft67CpzcRuXcHYTfRTLp3Rk';
-
 ////////////////////////////////////////////////////////////////////////
 //                          PACKAGE SEARCH                            //
 //                        FIX THE HTTP SEARCH                         //
@@ -75,10 +74,10 @@ async function packageUpload() {
     try {
         var localUrl = document.getElementById("url").value
         var content = document.getElementById("content").value
-        if (content === ""){
+        if (content === "") {
             content = null
         }
-        if (localUrl === ""){
+        if (localUrl === "") {
             localUrl = null
         }
         newurl = url + 'package/';
@@ -92,7 +91,7 @@ async function packageUpload() {
         })
         var items = await response.text()
         document.getElementById("return").innerHTML = "Upload Success.";
-        
+
     } catch (err) {
         document.getElementById("return").innerHTML = err;
     }
@@ -110,19 +109,19 @@ async function packageUpdate() {
         var ID = document.getElementById("ID").value
         var content = document.getElementById("content").value
         var URL = document.getElementById("URL").value
-        if (content === ""){
+        if (content === "") {
             content = null
         }
-        if (URL === ""){
+        if (URL === "") {
             URL = null
         }
-        if (name === ""){
+        if (name === "") {
             name = null
         }
-        if (ID === ""){
+        if (ID === "") {
             ID = null
         }
-        if (version === ""){
+        if (version === "") {
             version = null
         }
         newurl = url + 'package/' + document.getElementById('ID').value + '/';
@@ -132,13 +131,13 @@ async function packageUpdate() {
                 'X-Authorization': bearer,
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify({ 'metadata':{'ID': ID, 'Version': version, 'Name': name },'data':{'URL': URL, 'Content': content,}})
+            body: JSON.stringify({ 'metadata': { 'ID': ID, 'Version': version, 'Name': name }, 'data': { 'URL': URL, 'Content': content, } })
         })
-        if(response.status == 200){
+        if (response.status == 200) {
             var items = await response.text()
             document.getElementById("return").innerHTML = "Update Success.";
         }
-        else{
+        else {
             document.getElementById("return").innerHTML = "Update Failed. Ensure you have the proper fields filled out.";
         }
     } catch (err) {
@@ -152,16 +151,24 @@ async function packageUpdate() {
 ////////////////////////////////////////////////////////////////////////
 
 async function systemReset() {
-    try{
+    try {
         newurl = url + 'reset'
-        var response = await fetch(newurl, {
-            method: 'DELETE',
-            headers: {
-                'X-Authorization': bearer,
-            },
-        })
-        var items = await response.text()
-        document.getElementById("return").innerHTML = newurl;
+        var locbearer = localStorage.getItem("key");
+        if (localStorage.getItem("flag") == 0) {
+            document.getElementById("return").innerHTML = "Please enter your username and password for access."
+        } else {
+            var response = await fetch(newurl, {
+                method: 'DELETE',
+                headers: {
+                    'X-Authorization': locbearer,
+                },
+            })
+            console.log(localStorage.getItem("flag"))
+            var items = await response.text()
+            document.getElementById("return").innerHTML = "Success!";
+            localStorage.setItem("flag", 0)
+            console.log(localStorage.getItem("flag"))
+        }
     } catch (err) {
         document.getElementById("return").innerHTML = err;
     }
@@ -242,10 +249,10 @@ async function packageDirect() {
     try {
         var version = document.getElementById("searchbar").value
         var name = document.getElementById("searchbarName").value
-        if (version === ""){
+        if (version === "") {
             version = null
         }
-        if (name === ""){
+        if (name === "") {
             name = null
         }
         newurl = url + 'packages/';
@@ -257,9 +264,10 @@ async function packageDirect() {
             },
             body: JSON.stringify({ 'Version': version, 'Name': name })
         })
+        console.log(response)
         var items = await response.text();
         document.getElementById("return").innerHTML = JSON.stringify(JSON.parse(items));
-        
+
     } catch (err) {
         document.getElementById("return").innerHTML = err;
     }
@@ -274,15 +282,12 @@ async function loginFunc() {
     try {
         var username = document.getElementById("username").value;
         var password = document.getElementById("password").value
-        var isAdmin = false
-        if (username === ""){
+        var isAdmin = true
+        if (username === "") {
             username = null
         }
-        if (password === ""){
+        if (password === "") {
             password = null
-        }
-        if (username == "ece30861defaultadminuser" && password == "correcthorsebatterystaple123(!__+@**(A'+\"'\"+'\"`;DROP TABLE packages;"){
-            isAdmin = true
         }
         newurl = url + 'authenticate/';
         var response = await fetch(newurl, {
@@ -291,15 +296,54 @@ async function loginFunc() {
                 'X-Authorization': bearer,
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify({ 'User':{'name': username, 'isAdmin': isAdmin},'Secret':{'password': password}})
+            body: JSON.stringify({ 'User': { 'name': username, 'isAdmin': isAdmin }, 'Secret': { 'password': password } })
         })
-        if(response.status == 200){
-            //bearer = await response.text()
-            document.getElementById("return").innerHTML = response.text();
+        var items = await response.text();
+        //bearer = await response.text()
+        console.log(bearer)
+        document.getElementById("return").innerHTML = "Authentication Success. New Token added";
+        localStorage.setItem("key", items);
+        localStorage.setItem("flag", 1)
+        console.log(localStorage.getItem("flag"))
+    } catch (err) {
+        document.getElementById("return").innerHTML = err;
+    }
+}
+
+
+////////////////////////////////////////////////////////////////////////
+//                          SIZE COST                                 //
+//                        FIX THE HTTP SEARCH                         //
+////////////////////////////////////////////////////////////////////////
+
+async function sizeCost() {
+
+    try {
+        var content = document.getElementById("content").value
+        var name = document.getElementById("name").value
+        var locURL = document.getElementById("url").value
+        /*if (content === ""){
+            content = null
         }
-        else{
-            document.getElementById("return").innerHTML = "Login Failed";
+        if (name === ""){
+            name = null
         }
+        if (locURL === ""){
+            locURL = null
+        }*/
+        newurl = url + 'packages/';
+        var response = await fetch(newurl, {
+            method: 'POST',
+            headers: {
+                'X-Authorization': bearer,
+                'Content-Type': 'application/json'
+            },
+            body: [{ 'Content': content, 'Name': name, 'URL': locURL },],
+        })
+        console.log(response)
+        var items = await response.text();
+        document.getElementById("return").innerHTML = JSON.stringify(JSON.parse(items));
+
     } catch (err) {
         document.getElementById("return").innerHTML = err;
     }
